Handle ini files better

+++ src/ComposerInstaller.php
@@ -31,15 +31,16 @@ class ComposerInstaller
                 return false;   // Silently Fail.
             }
         }
-        if (!file_exists($ymlPath.'GCWorld_ORM.yml')) {
+        if (!file_exists($ymlPath.'GCWorld_ORM.yml') && !file_exists($ymlPath.'GCWorld_ORM.ini')) {
             $example = file_get_contents($myDir.$ds.'..'.$ds.'config'.$ds.'config.example.yml');
             file_put_contents($ymlPath.'GCWorld_ORM.yml', $example);
         }
 
-        file_put_contents($myDir.$ds.'..'.$ds.'config'.$ds.'config.yml', Yaml::dump([
-            'config_path' => $ymlPath.'GCWorld_ORM.yml',
-        ], 4));
-
+        if(!file_exists($myDir.$ds.'..'.$ds.'config'.$ds.'config.ini')) {
+            file_put_contents($myDir.$ds.'..'.$ds.'config'.$ds.'config.yml', Yaml::dump([
+                'config_path' => $ymlPath.'GCWorld_ORM.yml',
+            ], 4));
+        }
         return true;
     }
 }
